+++ MOP/src/GameObjects/Vehicles/Satsuma.cs
@@ -301,6 +301,10 @@ namespace MOP
 
             if (Rules.instance.SpecialRules.ExperimentalOptimization)
                 key = transform.Find("Dashboard/Steering/steering_column2/Ignition/Keys/Key").gameObject;
+
+            // Fix for reg plates Z fighting.
+            GameObject.Find("bootlid(Clone)").transform.Find("RegPlateRear").gameObject.GetComponent<Renderer>().material.renderQueue = 3001;
+            GameObject.Find("bumper front(Clone)").transform.Find("RegPlateFront").gameObject.GetComponent<Renderer>().material.renderQueue = 3001;
         }
 
         /// <summary>